NVMe fix and disallow blank encryption passphrases
Creating partitions on NVMe drives appears to take place so quickly that when the script was attempting to access the new partition, the OS was not yet aware of it. This is fixed by waiting until the partition is visible to return its path (new lines 169-172).

Blank encryption passphrases are pointless, throw error on blank passphrase in case someone is just hitting the Enter key to select all the defaults for the preceding queries.